lets see

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: rde-kwaa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/23 14:23:45 by rde-kwaa          #+#    #+#             */
-/*   Updated: 2018/07/20 16:19:29 by rde-kwaa         ###   ########.fr       */
+/*   Updated: 2018/07/21 12:58:57 by rde-kwaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int					ft_hidenp(char c, char *s);
 char				*ft_htoa(unsigned long n);
 int					ft_isdigit(int c);
 char				*ft_itoa(int n);
-int					ft_len(int n);
+size_t				ft_len(int n);
 char				*ft_litoa(long long int n);
 char				*ft_lotoa(unsigned long long int n);
 char				*ft_lutoa(unsigned long long int n);
